requirements.txt
pymongo
flask
flask-cors
python-dotenv
request